+++ routes/api.php
@@ -22,11 +22,18 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::get('user', function (Request $request) {
         return $request->user();
     });
+
     Route::apiResource('sinks', SinkApiController::class)->only(['index', 'show', 'update']);
+    Route::controller(SinkApiController::class)->group(function () {
+        Route::get('sinks/{sink}/scan', 'scanLocalFiles');
+    });
+
     Route::apiResource('sinks.schemas', SinkSchemaApiController::class)->only(['index', 'show']);
+
     Route::apiResource('sinks.chunks', ChunkController::class)->only(['index', 'update']);
     Route::controller(ChunkController::class)->group(function () {
         Route::get('sinks/{sink}/chunks/{chunk}/download', 'download');
     });
+
     Route::apiResource('batch', BatchApiController::class)->only(['index', 'show', 'destroy']);
 });